Project restructure to move utility functions and files inside common folder within Django app
