+++ sustainabilityGame/common/campusCoordinates.json
