+++ sustainabilityGame/common/travelTypes.py
@@ -1,3 +1,4 @@
+
 from enum import Enum
 
 # figures from https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/901692/conversion-factors-2020-methodology.pdf
