+++ sustainabilityGame/common/utils.py
@@ -1,6 +1,24 @@
-from travelTypes import TravelType
-import requests
+
 import json
+import requests
+from travelTypes import TravelType
+
+
+# These are the functions for calculating how much CO2 has been saved by the journey being done sustainably
+# Function takes in distance traveled, and a number representing how they did it. web app will call this fuction.
+def distanceToCO2(distance: float, transport: TravelType|int):
+    if type(transport) is int:
+        match transport: 
+                case 0: 
+                        transport = TravelType.BUS 
+                case 1:   
+                        transport = TravelType.TRAIN
+                case 2: 
+                        transport = TravelType.BIKE 
+                case 3: 
+                        transport = TravelType.WALK 
+    co2_saved = distance * abs(transport - TravelType.CAR)
+    return co2_saved 
 
 
 def locationToDistance(origin_lat: float, origin_long: float, dest_lat: float, dest_long: float, transport: TravelType):
@@ -54,7 +72,3 @@ def locationToDistance(origin_lat: float, origin_long: float, dest_lat: float, d
         if result[0]['condition'] == "ROUTE_EXISTS":
             return result[0]['distanceMeters']
     return 0
-
-
-# Test function
-print(locationToDistance(50.727163, -3.535225, 50.736235, -3.534693, TravelType.BIKE))
